requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-Alembic==2.0.1
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-CLI==0.4.0
psycopg2==2.9.6
python-dotenv==1.0.0
Werkzeug==2.2.3
alembic==1.10.3
SQLAlchemy==2.0.9
marshmallow==3.19.0
marshmallow_sqlalchemy==0.29.0
flask-marshmallow==0.15.0
flask-jwt-extended==4.5.2
PyJWT==2.7.0
aiofiles==23.1.0
asgiref==3.7.2
async-timeout==4.0.2
redis==4.6.0
bidict==0.22.1
flask-socketio==5.3.4
python-engineio==4.4.1
python-socketio==5.8.0
simple-websocket==0.10.1
flask-mail==0.9.1
flask-pika==0.3.8
Celery==5.3.1
